temp save

+++ Althea/TensorAlgebra/Sparse/CustomTypes.cs
@@ -206,7 +206,7 @@ namespace Althea.TensorAlgebra.Sparse
 			if (format == 0)
 				return Span<SparseTensorFormat>.Empty;
 			int f = (int)format;
-			byte c = f.CountBitSet();
+			int c = f.CountBitSet();
 			if (result.Length < c)
 				throw new ArgumentException(Resources.Parameter.WrongSize, nameof(result));
 			result = result[..c]; c = 0;
